+++ spec/dummy/config/routes.rb
@@ -1,5 +1,7 @@
 Rails.application.routes.draw do
   mount Unmagic::Passkeys::Engine => "/unmagic-passkeys"
 
-  use_unmagic_passkeys
+  resource  :session,      only: %i[new create destroy]
+  resource  :registration, only: %i[new create]
+  resources :passkeys,     only: %i[index create destroy]
 end